+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # MiniReelinator
 
+## Beta13 (July 22, 2014)
 * *[Beta13.rc1]*
   * [FEATURE]: Support generating wordpress shortcode embeds
   * [FEATURE]: Styles for wordpress shortcode embeds added
+  * [FEATURE]: Add support for new MR Player embed
+* *[/Beta13.rc1]*
 
 ## Beta12 (July 17, 2014)
 * *[Beta12.rc3]*